feat(mon-pix): move mirage oidc routes
