+++ mon-pix/mirage/config.js
@@ -13,7 +13,8 @@ import postSessionParticipation from './routes/post-session-participation';
 import postExpiredPasswordUpdates from './routes/post-expired-password-updates';
 import loadAnswerRoutes from './routes/answers/index';
 import loadAssessmentRoutes from './routes/assessments/index';
-import loadAuthRoutes from './routes/auth/index';
+import loadAuthenticationRoutes from './routes/authentication';
+import loadOidcAuthenticationRoutes from './routes/authentication/oidc';
 import loadCampaignParticipations from './routes/campaign-participations/index';
 import loadCertificationCourseRoutes from './routes/certification-courses/index';
 import loadCourseRoutes from './routes/courses/index';
@@ -39,7 +40,8 @@ export default function () {
 
   loadAnswerRoutes(this);
   loadAssessmentRoutes(this);
-  loadAuthRoutes(this);
+  loadAuthenticationRoutes(this);
+  loadOidcAuthenticationRoutes(this);
   loadCampaignParticipations(this);
   loadCertificationCourseRoutes(this);
   loadCourseRoutes(this);
@@ -83,37 +85,4 @@ export default function () {
   this.get('/certification-candidates/:id/subscriptions', getCertificationCandidatesSubscriptions);
 
   this.get('/certification-courses/:id');
-
-  this.get('/oidc/authentication-url', (schema, request) => {
-    const redirectUri = request.queryParams.redirect_uri;
-    return {
-      redirectTarget: `https://oidc/connexion/oauth2/authorize?redirect_uri=${redirectUri}`,
-      state: 'a8a3344f-6d7c-469d-9f84-bdd791e04fdf',
-      nonce: '555c86fe-ed0a-4a80-80f3-45b1f7c2df8c',
-    };
-  });
-
-  this.post('/oidc/users', (schema) => {
-    const createdUser = schema.users.create({
-      firstName: 'Lloyd',
-      lastName: 'Cé',
-    });
-
-    return {
-      access_token:
-        'aaa.' +
-        btoa(
-          `{"user_id":${createdUser.id},"source":"pole_emploi_connect","identity_provider":"POLE_EMPLOI","iat":1545321469,"exp":4702193958}`
-        ) +
-        '.bbb',
-      user_id: createdUser.id,
-    };
-  });
-
-  this.get('/oidc/redirect-logout-url', () => {
-    return {
-      redirectLogoutUrl:
-        'http://identity_provider_base_url/deconnexion?id_token_hint=ID_TOKEN&redirect_uri=http%3A%2F%2Flocalhost.fr%3A4200%2Fconnexion',
-    };
-  });
 }
